refactor: clean up code and optimise docker
